pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "mediapipe>=0.10.21",
    "mido>=1.3.3",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pygame>=2.6.1",
]
